fb: Keyguard disable challenge drag down on slimpin/simpuk screen
On simpin, simpuck and as well account change screen it does not make
any sense to give the user the abillity to drag down the challenge to show
the widget maximized (neither if widgets are enabled or disabled)
Reason: simpin/simpuck widget is not shown due of security reasons
and on account change the next expected behaviour is that the user gets a new
setup. So it does not make in this case as well sense

Just check for current securityscreen selection to block this behaviour. As
well move the enable/disable code to a proper place

Change-Id: I0b16d1e590252b537d6c71c4aa9b133ed832be45

Conflicts:
	packages/Keyguard/src/com/android/keyguard/KeyguardHostView.java