fb: finally fix carrier lable shows on first boot sometimes
Issue:
- boot up
- directly open via fast QS pulldown QS
- sometimes carrier label is shown

Expected behaviour: should not show

Fix: Visibility is directly set when the settings observer is registered first time.
If no update was done on the carrier label before the user pulled QS down the first time
it showed up. Vary rare case that this can happen
but should be handled. To fix it do not set the visibility directly and call the
updateCarrierAndWifiLabelVisibility(boolean) method to handle the correct visibility.

As well last cleanup on notification shorcuts...move the check if it is
active or not to the visbility method itself. Way cleaner

And remove the previous attempt to fix the above described issue due
it did not fix it at all

Change-Id: I901beb987c21fe0cdb7a5d9f16f850f4297c15f4

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java